requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MySQL==1.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.6.0
pylint==2.4.4
PyMySQL==0.9.3
python-dotenv==0.10.3
regex==2019.12.9
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
